Add a new class Serializable.
This is a base class from which other classes must inherit to make them
serializable. This class provides two main methods: serialize() and
deserialize(...) which return a serialized string representation of the object
and the object from it's serialized string representation respectively.

This class provides "true" serialization and is able to serialize dynamically
constructed objects. In other words, it serializes attributes added to live
objects and also serializes methods etc. and reconstructs the complete live
object.